refactor logic for handling non-wildcard redirects
This moves the logic for non-wildcard redirects (that is, simple path to
destination mappings) up into the gum.Server, rather than requiring each
handler to handle it directly.  Currently only the jekyll.Handler was
doing this.  Instead, the gum.Handler interface adds a URLs() method
that returns a map[string]string for the gum.Server to handle.

In moving this logic, it means that each gum.Handler does not
necessarily need to be an http.Handler.  In the case of handlers that
just return a mapping of short URLs to destinations, they can just hand
that map to the server and be done.

This has the added benefit for jekyll.Handler specifically in that it no
longer needs a prefix for URLs it handles.  It hands all of its
redirects, regardless of prefix, to the server and they get handled
appropriately.

Normal precedence rules for registering ServeMux handlers applies,
namely that longer patterns win.  Since the simple path to destination
mappings are registered on "/", any handler that is registered for an
explicit path prefix (like RedirectHandler) will take precedence.